requirements.txt
streamlit>=1.30.0
PyMuPDF>=1.23.0
google-generativeai>=0.3.2
python-dotenv>=1.0.1
PyPDF2>=3.0.1
pandas>=2.2.2
pyttsx3>=2.90
